Adding and Updating Files
Until here, it's possible to browse in the web site with a better appeareance because your styles was redone and i've created a new type called 'moviesDetails' to filter what is returned to us (clients), but it's not possible to do requests, it's returning an error 404 (not found) where i belive that is a problem with instances. In the future i'll solve these problems.